Remove redundant cast

+++ core/trino-spi/src/main/java/io/trino/spi/block/Int96ArrayBlock.java
@@ -125,7 +125,7 @@ public class Int96ArrayBlock
         if (valueIsNull != null) {
             consumer.accept(valueIsNull, sizeOf(valueIsNull));
         }
-        consumer.accept(this, (long) INSTANCE_SIZE);
+        consumer.accept(this, INSTANCE_SIZE);
     }
 
     @Override
